backend/requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
pillow
spacy
reportlab
python-multipart
